+++ uu-library-app/FormUI/Other-Operations/author-actions.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MySql.Data.MySqlClient;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -18,6 +19,7 @@ namespace uu_library_app
 {
     public partial class author_actions : Form
     {
+        MySqlConnection conn = new MySqlConnection("Server=172.21.54.3;uid=ASSEMSoft;pwd=Assemsoft1320..!;database=ASSEMSoft");
         AuthorManager manager = new AuthorManager(new AuthorDal());
 
         public author_actions()
@@ -25,18 +27,43 @@ namespace uu_library_app
             InitializeComponent();
         }
 
-     
+        private void listDataToTable()
+        {
+            DataTable dt = new DataTable();
+            MySqlDataAdapter da = new MySqlDataAdapter("Select * From Author WHERE deleted=false", conn);
+            da.Fill(dt);
+            dataGridView1.DataSource = dt;
+            dataGridView1.Columns[1].HeaderText = "";
+            dataGridView1.Columns[0].Visible = false;
+            dataGridView1.Columns[3].Visible = false;
+            dataGridView1.Columns[4].Visible = false;
+            dataGridView1.ColumnHeadersVisible = false;
+            dataGridView1.RowHeadersVisible = false;
+            dataGridView1.DefaultCellStyle.Font = new Font("Nirmala UI", 13);
+        }
+        private void clearAllFields()
+        {
+            txtId.Clear();
+            txtAd.Clear();
+            txtSoyad.Clear();
+        }
+
         private void btnEkle_Click(object sender, EventArgs e)
         {
-            string createGUID = System.Guid.NewGuid().ToString(); 
-            Author authorToAdd = new Author(createGUID, txtAd.Text, txtSoyad.Text);
+            string createGUID = System.Guid.NewGuid().ToString();
 
+            if (txtAd.Text == "" || txtSoyad.Text == "" || txtAd.Text.Length < 2 || txtSoyad.Text.Length < 2)
+            {
+                MessageBox.Show("Lütfen geçerli ve en az 2 karakter içeren değerler giriniz");
+                return;
+            }
+
+            Author authorToAdd = new Author(createGUID, txtAd.Text, txtSoyad.Text);
             try
             {
                 manager.Add(authorToAdd);
-                txtId.Clear();
-                txtAd.Clear();
-                txtSoyad.Clear();
+                listDataToTable();
+                clearAllFields();
             }
             catch (Exception)
             {
@@ -47,10 +74,17 @@ namespace uu_library_app
 
         private void btnSil_Click(object sender, EventArgs e)
         {
+            if(txtId.Text == "")
+            {
+                MessageBox.Show("Silmek istediğiniz konumu seçin!");
+                return;
+            }
             
             try
             {
                 manager.Delete(txtId.Text);
+                listDataToTable();
+                clearAllFields();
             }
             catch (Exception)
             {
@@ -62,16 +96,38 @@ namespace uu_library_app
 
         private void btnGuncelle_Click(object sender, EventArgs e)
         {
+            if(txtAd.Text == "" || txtSoyad.Text == "" || txtAd.Text.Length < 2 || txtSoyad.Text.Length < 2)
+            {
+                MessageBox.Show("Lütfen geçerli ve en az 2 karakter içeren değerler giriniz");
+                return;
+            }
+
             Author authorToUpdate = new Author(txtId.Text, txtAd.Text, txtSoyad.Text);
             try
             {
                 manager.Update(authorToUpdate);
+                listDataToTable();
+                clearAllFields();
             }
             catch (Exception)
             {
                 MessageBox.Show("Bir hata oluştu. Tekrar deneyiniz.");
                 throw;
             }
+        }
+
+        private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            txtId.Text = dataGridView1.Rows[e.RowIndex].Cells[0].Value.ToString();
+            txtAd.Text = dataGridView1.Rows[e.RowIndex].Cells[1].Value.ToString();
+            txtSoyad.Text = dataGridView1.Rows[e.RowIndex].Cells[2].Value.ToString();
+        }
+
+        private void author_actions_Load(object sender, EventArgs e)
+        {
+            conn.Open();
+            listDataToTable();
+            conn.Close();
         }
     }
 }
